Changed created by
Removed jc's name as he is not longer a councillor, nor the maintainer of the website, and the only this he changed was the theme. In my opinion, it takes more than that to be credited. I changed my status to created by, instead of maintained by, as I will no longer be personally maintaining or hosting the website, so it would be misleading not to change it. To whoever adopts this website, please do it proud, and don't just let it rot.